feat(expo): move rating prompt into onboarding paywall step
Triggers StoreReview on Continue from the "Free to use. Community
supported." screen instead of waiting until the feed has 3+ upcoming
events. Still one-shot via hasShownRatingPrompt.
